Refactor PEFConfigParams

+++ cmd_set_pef_config_params.go
@@ -22,7 +22,7 @@ func (req *SetPEFConfigParamsRequest) Pack() []byte {
 	out := make([]byte, 1+len(req.ConfigData))
 
 	// out[0] = req.ParamSelector
-	packUint8(req.ParamSelector, out, 0)
+	packUint8(uint8(req.ParamSelector), out, 0)
 	if len(req.ConfigData) > 0 {
 		packBytes(req.ConfigData, out, 1)
 	}
@@ -36,10 +36,8 @@ func (res *SetPEFConfigParamsResponse) Unpack(msg []byte) error {
 func (r *SetPEFConfigParamsResponse) CompletionCodes() map[uint8]string {
 	return map[uint8]string{
 		0x80: "parameter not supported",
-
 		// (This completion code provides a way to recognize that another party has already 'claimed' the parameters)"
 		0x81: "attempt to set the 'set in progress' value (in parameter #0) when not in the 'set complete' state.",
-
 		0x82: "attempt to write read-only parameter",
 		0x83: "attempt to read write-only parameter",
 	}
